ui next: Routing for tabbed pages
Refactor our Routes definition to allow for "tabbed" sub-pages: for example, the
"overview" and "graphs" tab on the Cluster page.

+ Adds route definitions for all "tabs" on the Nodes and Cluster page.
+ Add "Single Node" page, which does not yet have its tabs.
+ Removed all "Title" components; instead, routed components may statically
define a "Title" function, which the Layout component can find. This is the
approach suggested by react-router's "breadcrumbs" example.
+ Created "ListLink", a general purpose component to create an active
react-router Link for a List Element. This is more general than the "IconLink"
previously used by the navigation sidebar.
