+++ ui/next/app/components/layoutSidebar.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { Link, RouterOnContext } from "react-router";
+import { ListLink, LinkProps } from "./listLink";
 import * as Icons from "./icons";
 
 function trustIcon(iconSvg: string) {
@@ -7,55 +7,26 @@ function trustIcon(iconSvg: string) {
   return {__html: iconSvg};
 }
 
-interface IconLinkProps {
-  to: string;
-  className?: string;
+interface IconLinkProps extends LinkProps {
   icon?: string;
   title?: string;
-}
-
-interface IconLinkContext {
-  router: RouterOnContext;
 }
 
 /**
  * IconLink creats a react router Link which contains both a graphical icon and
  * a string title.
- *
- * TODO(mrtracy): During construction of 'next', we are explicitly avoiding any
- * CSS changes from the existing ui. After publishing 'next', we should
- * restructure the css of these links slightly so that the 'active' class is
- * present directly on the <a> element, instead of the containing <li> element;
- * that will allow us to use the react-router's <Link> class to provide the
- * activeClass, rather than the logic below.
  */
 class IconLink extends React.Component<IconLinkProps, {}> {
-  static contextTypes = {
-    router: React.PropTypes.object,
-  };
-
-  static defaultProps = {
-    className: "normal",
-  };
-
-  context: IconLinkContext;
-
   render() {
-    let {to, className, icon, title} = this.props;
-    if (this.context.router.isActive(to)) {
-      if (className) {
-        className += " active";
-      } else {
-        className = "active";
-      }
-    }
-    return <li className={className}>
-      <Link to={to}>
-        <div className=".image-container"
-             dangerouslySetInnerHTML={trustIcon(icon)}/>
-        <div>{title}</div>
-      </Link>
-    </li>;
+    let passProps = {
+      to: this.props.to,
+      className: this.props.className,
+    };
+    return <ListLink {...passProps} >
+      <div className=".image-container"
+           dangerouslySetInnerHTML={trustIcon(this.props.icon)}/>
+      <div>{this.props.title}</div>
+    </ListLink>;
   }
 }
 
@@ -69,7 +40,7 @@ export default class extends React.Component<{}, {}> {
     return <div id="header">
       <header>
         <ul className="nav">
-          <IconLink to="/" icon={Icons.cockroachIcon} className="cockroach"/>
+          <IconLink to="/" icon={Icons.cockroachIcon} className="cockroach" />
           <IconLink to="/cluster" icon={Icons.clusterIcon} title="Cluster"/>
           <IconLink to="/nodes" icon={Icons.nodesIcon} title="Nodes"/>
           <IconLink to="/databases" icon={Icons.databaseIcon} title="Databases"/>
