+++ ui/next/app/containers/clusterEvents.tsx
@@ -2,16 +2,16 @@
 import * as React from "react";
 
 /**
- * Renders the main content of the help us page.
+ * Renders the content of the Cluster events tab.
  */
 export default class extends React.Component<{}, {}> {
   static title() {
-    return <h2>Help Cockroach Labs</h2>;
+    return <h2>Cluster</h2>;
   }
 
   render() {
     return <div className="section">
-      <h1>HelpUs Page</h1>
+      <h1>Cluster Events</h1>
     </div>;
   }
 }
