+++ ui/next/app/containers/databases.tsx
@@ -2,21 +2,16 @@
 import * as React from "react";
 
 /**
- * DatabasesTitle renders the main content of the databases page.
+ * Renders the main content of the databases page.
  */
-export class DatabasesMain extends React.Component<{}, {}> {
+export default class extends React.Component<{}, {}> {
+  static title() {
+    return <h2>Databases</h2>;
+  }
+
   render() {
     return <div className="section">
       <h1>Databases Page</h1>
     </div>;
-  }
-}
-
-/**
- * DatabasesTitle renders the header of the databases page.
- */
-export class DatabasesTitle extends React.Component<{}, {}> {
-  render() {
-    return <h2>Databases</h2>;
   }
 }
