+++ ui/next/app/js/protos.js
@@ -443,7 +443,8 @@ module.exports = require("protobufjs").newBuilder({})['import']({
                             "name": "timestamp_nanos",
                             "id": 1,
                             "options": {
-                                "(gogoproto.nullable)": false
+                                "(gogoproto.nullable)": false,
+                                "(gogoproto.jsontag)": "timestamp_nanos,string"
                             }
                         },
                         {
@@ -548,7 +549,8 @@ module.exports = require("protobufjs").newBuilder({})['import']({
                             "name": "start_nanos",
                             "id": 1,
                             "options": {
-                                "(gogoproto.nullable)": false
+                                "(gogoproto.nullable)": false,
+                                "(gogoproto.jsontag)": "start_nanos,string"
                             }
                         },
                         {
@@ -557,7 +559,8 @@ module.exports = require("protobufjs").newBuilder({})['import']({
                             "name": "end_nanos",
                             "id": 2,
                             "options": {
-                                "(gogoproto.nullable)": false
+                                "(gogoproto.nullable)": false,
+                                "(gogoproto.jsontag)": "end_nanos,string"
                             }
                         },
                         {
